requirements.txt
Django==3.1.5
django-phonenumber-field==4.0.0
phonenumberslite==8.11.4
